Claw Rollers Collect, Release, Default Spin Commands
I cannot be in the shop today, so I finished up the stuff I was working on last time. Someone should deploy this code and make sure that the claw can collect a game piece, release a game piece, and hold a cube piece. Also, currently, the game piece can only be changed through the code. We need to implement a way to change the game piece through shuffleboard.